+++ src/schemas/image.schema.ts
@@ -2,6 +2,6 @@ import {z} from "zod";
 
 export const ImageIdSchema = z.object({
     params: z.object({
-        id: z.coerce.number()
+        id: z.coerce.number().int()
     })
 })
